pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sp-rag"
version = "0.1.0"
description = "RAG pipeline for Russian construction norms (SNiP, SP)"
readme = "README.md"
authors = [{ name = "wedwincode", email = "edwindenisovelers9864@gmail.com" }]
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4==4.14.2",
    "chromadb-client==1.1.1",
    "html-to-markdown==1.10.0",
    "matplotlib==3.9.0",
    "pandas==2.3.3",
    "pydantic==2.12.3",
    "pydantic-settings==2.11.0",
    "requests==2.32.5",
    "yandex-cloud-ml-sdk==0.16.0",
]


[tool.setuptools.packages.find]
where = ["src"]
